requirements.txt
tortoise-orm~=0.18.1
jishaku~=2.3.2
psutil~=5.9.0
git+https://github.com/Rapptz/discord.py@master
aerich~=0.6.2
